requirements.txt
requests
PyYAML
pytest
pytest-html
Faker
behave
pytest-bdd
pytest-xdist